requirements.txt
accelerate==1.10.1
certifi==2025.8.3
charset-normalizer==3.4.3
coloredlogs==15.0.1
contourpy==1.3.3
cycler==0.12.1
decord==0.6.0
einops==0.8.1
filelock==3.19.1
flatbuffers==25.2.10
fonttools==4.60.0
fsspec==2025.9.0
ftfy==6.3.1
hf-xet==1.1.9
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
ml_dtypes==0.5.3
modelscope==1.29.2
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-ml-py==13.580.65
nvidia-nccl-cu12==2.27.3
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvtx-cu12==12.8.90
onnx==1.19.0
onnxruntime==1.22.1
open_clip_torch==3.2.0
opencv-python==4.12.0.88
pandas==2.3.2
peft==0.17.1
pillow==11.3.0
polars==1.33.1
protobuf==6.32.0
pycocotools==2.0.10
pynvml==13.0.1
pyparsing==3.2.4
pytz==2025.2
PyYAML==6.0.2
regex==2025.9.1
requests==2.32.5
safetensors==0.6.2
scipy==1.16.2
sentencepiece==0.2.1
sympy==1.14.0
timm==1.0.20
tokenizers==0.22.0
torch==2.8.0
torchvision==0.23.0
tqdm==4.67.1
transformers==4.56.1
triton==3.4.0
tzdata==2025.2
ultralytics==8.3.201
ultralytics-thop==2.0.17
urllib3==2.5.0